Moved server to sub module
