+++ ahnlich/db/Cargo.toml
@@ -7,9 +7,6 @@ edition = "2021"
 name = "ahnlich-db"
 path = "src/main.rs"
 
-[lib]
-path = "src/lib.rs"
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
